test: posts.view.py coverage complete
